moved config to a package
